added Web-Worker for QoS as QoS-processor
