+++ package.json
@@ -44,6 +44,7 @@
     "bootstrap": "4.5.2",
     "chart.js": "2.9.3",
     "chartjs-plugin-datalabels": "0.7.0",
+    "chartjs-plugin-responsive-downsample": "^1.1.3",
     "component": "1.1.0",
     "core-js": "3.6.5",
     "d3": "6.1.1",
