+++ src/app/components/search-grid-call/detail-dialog/detail-dialog.component.ts
@@ -20,8 +20,16 @@ import { ChangeDetectorRef } from '@angular/core';
     changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class DetailDialogComponent implements OnInit {
+    _sipDataItem: any;
     @Input() titleId: string;
-    @Input() sipDataItem: any;
+    @Input() set sipDataItem(val: any) {
+        this._sipDataItem = val;
+        this.changeDetectorRefs.detectChanges();
+        console.log('sipDataItem::', performance.now(), val);
+    }
+    get sipDataItem() {
+        return this._sipDataItem;
+    }
 
     @Input() headerColor: any;
     @Input() mouseEventData: any;
@@ -76,29 +84,29 @@ export class DetailDialogComponent implements OnInit {
             this.checkStatusTabs();
 
             const filterByParam = param => Object.keys(
-                    this.sipDataItem.data.messages
-                        .map(i => i[param])
-                        .reduce((a, b) => (a[b] = 1, a), {})
-                ).map((i: any) => {
-                    const obj = {
-                        selected: true,
-                        title: (param === 'payloadType' ? Functions.methodCheck(null, 1 * i) : i)
-                    };
-                    obj[param] = i;
-                    return obj;
-                });
+                this.sipDataItem.data.messages
+                    .map(i => i[param])
+                    .reduce((a, b) => (a[b] = 1, a), {})
+            ).map((i: any) => {
+                const obj = {
+                    selected: true,
+                    title: (param === 'payloadType' ? Functions.methodCheck(null, 1 * i) : i)
+                };
+                obj[param] = i;
+                return obj;
+            });
 
             this.checkboxListFilterPayloadType = filterByParam('payloadType');
             this.setFiltersByAdvanced();
             const ports = [].concat(filterByParam('dstPort'), filterByParam('srcPort'));
             this.checkboxListFilterPort = Object.keys(ports
-                    .map(i => i.title )
-                    .reduce((a, b) => (a[b] = a[b] ? a[b] + 1 : 1, a), {})
-                ).map(i => ({
-                    selected: true,
-                    title: i,
-                    port: i
-                }));
+                .map(i => i.title)
+                .reduce((a, b) => (a[b] = a[b] ? a[b] + 1 : 1, a), {})
+            ).map(i => ({
+                selected: true,
+                title: i,
+                port: i
+            }));
 
             this.checkboxListFilterCallId = filterByParam('sid');
             this.changeDetectorRefs.detectChanges();
@@ -117,7 +125,7 @@ export class DetailDialogComponent implements OnInit {
 
     @Output() openMessage: EventEmitter<any> = new EventEmitter();
     @Output() close: EventEmitter<any> = new EventEmitter();
-    @ViewChild('filterContainer', {static: false}) filterContainer: ElementRef;
+    @ViewChild('filterContainer', { static: false }) filterContainer: ElementRef;
     dataLogs: Array<any>;
 
     constructor(
@@ -125,7 +133,7 @@ export class DetailDialogComponent implements OnInit {
         private changeDetectorRefs: ChangeDetectorRef
     ) { }
 
-    ngOnInit () {
+    ngOnInit() {
         this.setTabByAdvanced();
         if (this.sipDataItem) {
             this.dataLogs = this.sipDataItem.data.messages.filter(i => !i.method).map(i => ({ payload: i }));
@@ -156,7 +164,7 @@ export class DetailDialogComponent implements OnInit {
             }
         });
     }
-    onClose () {
+    onClose() {
         this.close.emit();
         this.changeDetectorRefs.detectChanges();
     }
@@ -178,7 +186,7 @@ export class DetailDialogComponent implements OnInit {
         this.changeDetectorRefs.detectChanges();
     }
 
-    onBrowserWindow (event) {
+    onBrowserWindow(event) {
         this.isBrowserWindow = event;
         this.changeDetectorRefs.detectChanges();
     }
@@ -192,7 +200,7 @@ export class DetailDialogComponent implements OnInit {
                     if (setting && setting[0] && setting[0].data) {
                         const filters = setting[0].data.tabfilterconfig;
                         const filterBackup = Functions.cloneObject(this.checkboxListFilterPayloadType);
-                        if (this.checkboxListFilterPayloadType.length > 1 ) {
+                        if (this.checkboxListFilterPayloadType.length > 1) {
                             this.checkboxListFilterPayloadType.forEach(filter => {
                                 filter.selected = filters[filter.title];
                             });
@@ -242,13 +250,13 @@ export class DetailDialogComponent implements OnInit {
         }
         if (this.combineType === '1none') {
             this.isCombineByAlias = false;
-            this.isSimplifyPort   = false;
+            this.isSimplifyPort = false;
         } else if (this.combineType === '2alias') {
             this.isCombineByAlias = true;
-            this.isSimplifyPort   = true;
+            this.isSimplifyPort = true;
         } else if (this.combineType === '3port') {
             this.isCombineByAlias = false;
-            this.isSimplifyPort   = true;
+            this.isSimplifyPort = true;
         }
         setTimeout(() => {
             const fc = Functions.cloneObject;
