+++ src/app/components/search-grid-call/detail-dialog/tab-qos/tab-qos.component.html
@@ -22,7 +22,7 @@
           </div>
         </div>
       </div>
-      <!-- [colors]="lineChartColorsRTP" -->
+    
     </div>
     
 
@@ -91,7 +91,7 @@
           </div>
         </div>
       </div>
-      <!-- [colors]="lineChartColors" -->
+    
     </div>
