+++ src/environments/environment.dev.ts
@@ -3,6 +3,6 @@ import { VERSION } from '../VERSION';
 export const environment = {
   production: false,
   environment: VERSION + '(dev)',
-  apiUrl: window.location.protocol + '//' + (window.location.host) + '/api/v3'
+    apiUrl: 'http://de9.sipcapture.io/api/v3'
 
 };
